take above comments into account
